improvments, improvments and IMPROOOOVMÄNTZ!

+++ C#/Project_Sahlgrenska/Project_Sahlgrenska/SearchPatient.xaml.cs
@@ -30,6 +30,7 @@ namespace Project_Sahlgrenska
 
         private void Button_ClickSearch(object sender, RoutedEventArgs e)
         {
+            patientCritical.Text = "";
             UpdateSearch();
         }
 
@@ -86,12 +87,12 @@ namespace Project_Sahlgrenska
                 patientRoom.Text = list[11];
                 patientEquipment.Text = list[17];
                 patientHistory.Text = list[6];
+                string critical = list[7].ToUpper();
 
-                if (list[7] == "Yes")
+                if (critical.Contains("Y"))
                 {
                     patientCritical.Text = "PATIENT I KRITISKT LÄGE";
                 }
-
             }
             else
             {
@@ -108,11 +109,17 @@ namespace Project_Sahlgrenska
 
         private void Button_SentHome(object sender, RoutedEventArgs e)
         {
-            Bot.Update("UPDATE bt0mlsay6vs1xbceqzzn.patients SET History = CONCAT(NOW(), ' ++ Patienten utskriven från sjukhuset och " + patientRoom.Text + "\n\', COALESCE(CONCAT(CHAR(10), History), '')), SentHome = NOW() WHERE ID = '" + patientId.Text + "';");
-            Bot.Update("DELETE FROM patients_has_rooms WHERE rooms_ID = '" + patientRoom.Text + "';");
-            Bot.Update("SET SQL_SAFE_UPDATES = 0; UPDATE rooms SET beds = beds + 1 WHERE id = '" + patientRoom.Text + "'; SET SQL_SAFE_UPDATES = 1;");
-            Bot.Update("call update_vaccant_yes()");
-            UpdateSearch();
+            if (patientRoom.Text == "")
+            {
+                Bot.Update("UPDATE bt0mlsay6vs1xbceqzzn.patients SET History = CONCAT(NOW(), ' ++ Patienten utskriven från sjukhuset."  + "\n\', COALESCE(CONCAT(CHAR(10), History), '')), SentHome = NOW() WHERE ID = '" + patientId.Text + "';");
+                UpdateSearch();
+            }
+            else
+                Bot.Update("UPDATE bt0mlsay6vs1xbceqzzn.patients SET History = CONCAT(NOW(), ' ++ Patienten utskriven från sjukhuset och rum:  " + patientRoom.Text + ".\n\', COALESCE(CONCAT(CHAR(10), History), '')), SentHome = NOW() WHERE ID = '" + patientId.Text + "';");
+                Bot.Update("DELETE FROM patients_has_rooms WHERE rooms_ID = '" + patientRoom.Text + "';");
+                Bot.Update("SET SQL_SAFE_UPDATES = 0; UPDATE rooms SET beds = beds + 1 WHERE id = '" + patientRoom.Text + "'; SET SQL_SAFE_UPDATES = 1;");
+                Bot.Update("call update_vaccant_yes()");
+                UpdateSearch();
         }
 
         private void updateJournal_TextChanged(object sender, System.Windows.Controls.TextChangedEventArgs e)
